refactor: move screen step types (#9)
