--- packages/ui/src/components/pages/SetupWallet/types.ts
@@ -1,11 +1,0 @@
-export enum NewWalletScreenStep {
-  ChooseWalletPassword,
-  ConfirmWalletPassword,
-  BackupSecretRecoveryPhrase,
-}
-
-export enum RestoreWalletScreenStep {
-  EnterSecretRecoveryPhrase,
-  ChooseWalletPassword,
-  ConfirmWalletPassword,
-}
