+++ packages/ui/src/types.ts
@@ -31,3 +31,15 @@ export enum SettingsDialogScreen {
   SettingsWallet,
   BackupSecretPhrase,
 }
+
+export enum NewWalletScreenStep {
+  ChooseWalletPassword,
+  ConfirmWalletPassword,
+  BackupSecretRecoveryPhrase,
+}
+
+export enum RestoreWalletScreenStep {
+  EnterSecretRecoveryPhrase,
+  ChooseWalletPassword,
+  ConfirmWalletPassword,
+}
